moved mock-data into its own directory
