+++ src/applications/health-care-questionnaire/shared/api/load-questionnaires.js
@@ -9,7 +9,7 @@ const loadQuestionnaires = async () => {
   if (USE_MOCK_DATA) {
     promise = new Promise(resolve => {
       setTimeout(() => {
-        import(/* webpackChunkName: "my-questionnaires-sample-json" */ './my-questionnaires.sample.json').then(
+        import(/* webpackChunkName: "my-questionnaires-sample-json" */ './mock-data/my-questionnaires.sample.json').then(
           module => {
             resolve(module.default);
           },
